Move common utils to core, update wells imports
