+++ projects/subgraph-core/utils/Constants.ts
