+++ projects/subgraph-wells/tests/Aquifer.test.ts
@@ -1,31 +1,31 @@
 import { afterEach, assert, beforeEach, clearStore, describe, test } from "matchstick-as/assembly/index";
-import { BEAN_ERC20, WETH } from "../src/utils/Constants";
+import { BEAN_ERC20, WETH } from "../../subgraph-core/utils/Constants";
 import { boreDefaultWell } from "./helpers/Aquifer";
 import { AQUIFER, PUMP, WELL } from "./helpers/Constants";
 
 describe("Aquifer Well Deployment", () => {
-    beforeEach(() => {
-        boreDefaultWell()
-    })
+  beforeEach(() => {
+    boreDefaultWell();
+  });
 
-    afterEach(() => {
-        clearStore()
-    })
+  afterEach(() => {
+    clearStore();
+  });
 
-    test("Aquifer entity exists", () => {
-        assert.fieldEquals("Aquifer", AQUIFER.toHexString(), "id", AQUIFER.toHexString())
-    })
+  test("Aquifer entity exists", () => {
+    assert.fieldEquals("Aquifer", AQUIFER.toHexString(), "id", AQUIFER.toHexString());
+  });
 
-    test("Well entity exists", () => {
-        assert.fieldEquals("Well", WELL.toHexString(), "id", WELL.toHexString())
-    })
+  test("Well entity exists", () => {
+    assert.fieldEquals("Well", WELL.toHexString(), "id", WELL.toHexString());
+  });
 
-    test("Token entities exists", () => {
-        assert.fieldEquals("Token", BEAN_ERC20.toHexString(), "id", BEAN_ERC20.toHexString())
-        assert.fieldEquals("Token", WETH.toHexString(), "id", WETH.toHexString())
-    })
+  test("Token entities exists", () => {
+    assert.fieldEquals("Token", BEAN_ERC20.toHexString(), "id", BEAN_ERC20.toHexString());
+    assert.fieldEquals("Token", WETH.toHexString(), "id", WETH.toHexString());
+  });
 
-    test("Pump entity exists", () => {
-        assert.fieldEquals("Pump", PUMP.toHexString() + '-' + WELL.toHexString(), "id", PUMP.toHexString() + '-' + WELL.toHexString())
-    })
-})
+  test("Pump entity exists", () => {
+    assert.fieldEquals("Pump", PUMP.toHexString() + "-" + WELL.toHexString(), "id", PUMP.toHexString() + "-" + WELL.toHexString());
+  });
+});
