+++ projects/subgraph-wells/tests/helpers/Aquifer.ts
@@ -2,7 +2,7 @@ import { Address, Bytes, ethereum } from "@graphprotocol/graph-ts";
 import { newMockEvent } from "matchstick-as/assembly/index";
 import { BoreWell } from "../../generated/Aquifer/Aquifer";
 import { handleBoreWell } from "../../src/templates/AquiferHandler";
-import { BEAN_ERC20, WETH } from "../../src/utils/Constants";
+import { BEAN_ERC20, WETH } from "../../../subgraph-core/utils/Constants";
 import { AQUIFER, IMPLEMENTATION, PUMP, WELL, WELL_DATA, WELL_FUNCTION } from "./Constants";
 import { createContractCallMocks } from "./Functions";
